Revert "Added ability to send consistent global messages to users"
This reverts commit 640b96faa2cb4869e1f2097d4947df9e8a239150.

+++ src/GameState/StateValueNumber.cpp
@@ -11,7 +11,7 @@ StateValueNumber::StateValueNumber(double val) :
         valueType(NUMBER)
 {}
 
-int& StateValueNumber::getValue() {
+double& StateValueNumber::getValue() {
     return value;
 }
